Break apart CubeXr's main.cpp
