+++ projects/04_cube_xr/CubeXr.h
@@ -1,0 +1,54 @@
+// Copyright 2023 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "ppx/math_config.h"
+#include "ppx/ppx.h"
+using namespace ppx;
+
+class CubeXrApp 
+    : public ppx::Application
+{
+public:
+    virtual void Config(ppx::ApplicationSettings& settings) override;
+    virtual void Setup() override;
+    virtual void Render() override;
+
+private:
+    struct PerFrame
+    {
+        ppx::grfx::CommandBufferPtr cmd;
+        ppx::grfx::SemaphorePtr     imageAcquiredSemaphore;
+        ppx::grfx::FencePtr         imageAcquiredFence;
+        ppx::grfx::SemaphorePtr     renderCompleteSemaphore;
+        ppx::grfx::FencePtr         renderCompleteFence;
+
+        // XR UI per frame elements.
+        ppx::grfx::CommandBufferPtr uiCmd;
+        ppx::grfx::FencePtr         uiRenderCompleteFence;
+    };
+
+    std::vector<PerFrame>             mPerFrame;
+    ppx::grfx::ShaderModulePtr        mVS;
+    ppx::grfx::ShaderModulePtr        mPS;
+    ppx::grfx::PipelineInterfacePtr   mPipelineInterface;
+    ppx::grfx::GraphicsPipelinePtr    mPipeline;
+    ppx::grfx::BufferPtr              mVertexBuffer;
+    ppx::grfx::DescriptorPoolPtr      mDescriptorPool;
+    ppx::grfx::DescriptorSetLayoutPtr mDescriptorSetLayout;
+    ppx::grfx::DescriptorSetPtr       mDescriptorSet;
+    ppx::grfx::BufferPtr              mUniformBuffer;
+    grfx::Viewport                    mViewport;
+    grfx::Rect                        mScissorRect;
+    grfx::VertexBinding               mVertexBinding;
+};
